+++ image_transport/test/test_single_subscriber_publisher.cpp
@@ -60,6 +60,10 @@ protected:
     node.reset();
   }
 
+  static void TearDownTestSuite()
+  {
+    rclcpp::shutdown();
+  }
 
   rclcpp::Node::SharedPtr node;
 };